+++ web-marketplace/src/components/organisms/BuyCreditsModal/BuyCreditsModal.SetSelectedSellOrderElement.tsx
@@ -3,7 +3,7 @@ import { useFormikContext } from 'formik';
 
 import { UseStateSetter } from 'types/react/use-state';
 
-import { UISellOrderInfo } from 'pages/Projects/Projects.types';
+import { UISellOrderInfo } from 'pages/Projects/AllProjects/AllProjects.types';
 
 import { BuyCreditsValues } from './BuyCreditsModal';
