+++ web-marketplace/src/hooks/projects/useProjectsWithOrders.ts
@@ -18,13 +18,13 @@ import { getAllSanityCreditClassesQuery } from 'lib/queries/react-query/sanity/g
 import { getAllSanityPrefinanceProjectsQuery } from 'lib/queries/react-query/sanity/getAllPrefinanceProjectsQuery/getAllPrefinanceProjectsQuery';
 import { useWallet } from 'lib/wallet/wallet';
 
-import { useFetchAllOffChainProjects } from 'pages/Projects/hooks/useOffChainProjects';
-import { ProjectsSellOrders } from 'pages/Projects/hooks/useProjectsSellOrders.types';
-import { UNREGISTERED_PATH } from 'pages/Projects/Projects.constants';
+import { UNREGISTERED_PATH } from 'pages/Projects/AllProjects/AllProjects.constants';
+import { useFetchAllOffChainProjects } from 'pages/Projects/AllProjects/hooks/useOffChainProjects';
+import { ProjectsSellOrders } from 'pages/Projects/AllProjects/hooks/useProjectsSellOrders.types';
 import {
   sortPinnedProject,
   sortProjects,
-} from 'pages/Projects/utils/sortProjects';
+} from 'pages/Projects/AllProjects/utils/sortProjects';
 import { useClassesWithMetadata } from 'hooks/classes/useClassesWithMetadata';
 
 import { useLastRandomProjects } from './useLastRandomProjects';
