+++ web-marketplace/src/pages/Marketplace/Storefront/hooks/useBuySellOrderSubmit.tsx
@@ -17,7 +17,7 @@ import { microToDenom } from 'lib/denom.utils';
 import { BuyFailureEvent, BuySuccessEvent } from 'lib/tracker/types';
 import { useTracker } from 'lib/tracker/useTracker';
 
-import { normalizeToUISellOrderInfo } from 'pages/Projects/hooks/useProjectsSellOrders.utils';
+import { normalizeToUISellOrderInfo } from 'pages/Projects/AllProjects/hooks/useProjectsSellOrders.utils';
 import DenomIcon from 'components/molecules/DenomIcon';
 import { BuyCreditsValues } from 'components/organisms';
 import { SignAndBroadcastType } from 'hooks/useMsgClient';
