+++ web-marketplace/src/routes.tsx
@@ -34,7 +34,7 @@ import { homeLoader } from 'pages/Home/Home.loader';
 import { storefrontLoader } from 'pages/Marketplace/Storefront/Storefront.loader';
 import { ProfileEditMain } from 'pages/ProfileEdit/ProfileEdit.Main';
 import { ProfileEditSettings } from 'pages/ProfileEdit/ProfileEdit.Settings';
-import { projectsLoader } from 'pages/Projects/Projects.loader';
+import { projectsLoader } from 'pages/Projects/AllProjects/AllProjects.loader';
 import Settings from 'pages/Settings';
 import { AuthRoute } from 'components/atoms/AuthRoute';
 import { KeplrOrAuthRoute } from 'components/atoms/KeplrOrAuthRoute';
@@ -46,6 +46,7 @@ import { KeplrRoute } from './components/atoms';
 import { ProjectMetadata } from './pages/ProjectMetadata/ProjectMetadata';
 
 const Additionality = lazy(() => import('./pages/Additionality'));
+const AllProjects = lazy(() => import('./pages/Projects/AllProjects'));
 const BasicInfo = lazy(() => import('./pages/BasicInfo'));
 const BatchDetails = lazy(() => import('./pages/BatchDetails'));
 const BasketDetails = lazy(() => import('./pages/BasketDetails'));
@@ -67,7 +68,7 @@ const Media = lazy(() => import('./pages/Media'));
 const MethodologyDetails = lazy(() => import('./pages/MethodologyDetails'));
 const NotFoundPage = lazy(() => import('./pages/NotFound'));
 const Project = lazy(() => import('./pages/Project'));
-const Projects = lazy(() => import('./pages/Projects'));
+const Projects = lazy(() => import('./pages/Projects/AllProjects'));
 const ProjectCreate = lazy(() => import('./pages/ProjectCreate'));
 const ProjectFinished = lazy(() => import('./pages/ProjectFinished'));
 const ProjectLocation = lazy(() => import('./pages/ProjectLocation'));
@@ -140,13 +141,16 @@ export const getRoutes = ({
           path="methodology-review-process"
           element={<Navigate to="/" replace />}
         />
-        <Route
-          path="projects/:page"
-          element={<Projects />}
-          loader={projectsLoader({
-            queryClient: reactQueryClient,
-          })}
-        />
+        <Route path="projects" element={<Projects />}>
+          <Route index element={<Navigate to="1" />} />
+          <Route
+            path=":page"
+            element={<AllProjects />}
+            loader={projectsLoader({
+              queryClient: reactQueryClient,
+            })}
+          />
+        </Route>
         <Route
           path="project/:projectId"
           element={<Project />}
